fix: refactor: create canonical folder
