+++ src/rules/canonical-vocabulary/utils.ts
@@ -1,0 +1,13 @@
+import { RuleOption } from '../../utils/types';
+
+export const getMessageByParams = (options: RuleOption): string => {
+  if (options?.message) {
+    return options.message;
+  }
+
+  if (options.fixTo) {
+    return 'The term <word> is not recommended, use the term <fixTo>';
+  }
+
+  return 'The term <word> is not recommended';
+};
